#477 Relocate `Enableable` out of dedicated package
