+++ hartshorn-persistence/src/main/java/org/dockbox/hartshorn/persistence/hibernate/HibernateJpaRepository.java
@@ -24,7 +24,7 @@ import org.dockbox.hartshorn.core.context.ApplicationContext;
 import org.dockbox.hartshorn.core.context.element.TypeContext;
 import org.dockbox.hartshorn.core.domain.Exceptional;
 import org.dockbox.hartshorn.core.exceptions.ApplicationException;
-import org.dockbox.hartshorn.core.properties.Enableable;
+import org.dockbox.hartshorn.core.Enableable;
 import org.dockbox.hartshorn.persistence.JpaRepository;
 import org.dockbox.hartshorn.persistence.context.EntityContext;
 import org.dockbox.hartshorn.persistence.properties.PersistenceConnection;
